refactor: QueryBuilder 인터페이스화
